+++ src/main/java/com/bizcub/hideItemFrame/platforms/forge/Forge.java
@@ -1,8 +1,9 @@
-//? if forge {
+//? forge {
 /*package com.bizcub.hideItemFrame.platforms.forge;
 
 import net.minecraftforge.fml.common.Mod;
 
-@Mod("hide_item_frame")
-public class Forge {}
-*///?}
+@Mod(Forge.modId)
+public class Forge {
+    public static final String modId = /^$ mod_id^/ "hide_item_frame";
+}*///?}